task 3.1.4 is done

+++ src/main/java/ru/kata/spring/boot_security/demo/controllers/AdminController.java
@@ -2,21 +2,20 @@ package ru.kata.spring.boot_security.demo.controllers;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import ru.kata.spring.boot_security.demo.exception_handling.NoSuchUserException;
 import ru.kata.spring.boot_security.demo.models.Role;
 import ru.kata.spring.boot_security.demo.models.User;
 import ru.kata.spring.boot_security.demo.repositories.RoleRepository;
 import ru.kata.spring.boot_security.demo.services.UserService;
 
-import java.security.Principal;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
-@Controller
+@RestController
 @RequestMapping("/admin")
+@CrossOrigin(origins = "http://127.0.0.1:5500")
 public class AdminController {
 
     @Autowired
@@ -26,57 +25,48 @@ public class AdminController {
     private RoleRepository roleRepository;
 
     @GetMapping
-    public String showAllUsers(Model model, Principal principal) {
-        List<User> allUsers = userService.showAllUsers();
-        User user = userService.findUserByUsername(principal.getName());
-        List<Role> roles = roleRepository.findAll();
-        model.addAttribute("allUsers", allUsers);
-        model.addAttribute("user", user);
-        model.addAttribute("roles", roles);
-        return "allUsers";
+    public ResponseEntity<List<User>> showAllUsers() {
+        return ResponseEntity.ok(userService.showAllUsers());
     }
 
-    @GetMapping("/newUser")
-    public String newUser(Model model) {
-        List<Role> roles = roleRepository.findAll();
-        model.addAttribute("user", new User());
-        model.addAttribute("roles", roles);
-        return "newUser";
-    }
 
-    @PostMapping("/addUser")
-    public String addUser(@ModelAttribute("user") User user, @RequestParam(value = "roles", required = false) Set<Integer> roleIds) {
-        if (roleIds != null) {
-            Set<Role> roles = roleRepository.findAll().stream()
-                    .filter(role -> roleIds.contains(role.getId()))
-                    .collect(Collectors.toSet());
-            user.setRoles(roles);
-        }
-        userService.saveUser(user);
-        return "redirect:/admin";
-    }
-
-    @GetMapping("/editUser")
-    public String editUser(@RequestParam("id") Integer id, Model model) {
+    @GetMapping("/{id}")
+    public ResponseEntity<User> getUser(@PathVariable Integer id) {
         User user = userService.showUser(id);
-        List<Role> roles = roleRepository.findAll();
-        model.addAttribute("user", user);
-        model.addAttribute("roles", roles);
-        return "editUser";
+        if (user == null) {
+            throw new NoSuchUserException("User with ID = " + id + " not found");
+        }
+        return ResponseEntity.ok(user);
     }
 
-    @PostMapping("/updateUser")
-    public String updateUser(@ModelAttribute("user") User user, @RequestParam("roles") Set<Integer> roles) {
+    @PostMapping
+    public ResponseEntity<HttpStatus> addNewUser(@RequestBody User user) {
+        userService.saveUser(user);
+        return ResponseEntity.ok(HttpStatus.CREATED);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<User> updateUser(@PathVariable Integer id, @RequestBody User user) {
+        User existingUser = userService.showUser(id);
+        if (existingUser == null) {
+            throw new NoSuchUserException("User with ID = " + id + " not found");
+        }
         userService.updateUser(user);
-        return "redirect:/admin";
+        return ResponseEntity.ok(user);
     }
 
-    @PostMapping("/deleteUser")
-    public String deleteUser(@RequestParam("id") int id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<HttpStatus> deleteUser(@PathVariable Integer id) {
+        User user = userService.showUser(id);
+        if (user == null) {
+            throw new NoSuchUserException("User with ID = " + id + " not found");
+        }
         userService.deleteUser(id);
-        return "redirect:/admin";
-
-
+        return ResponseEntity.ok().build();
     }
 
+    @GetMapping("/roles")
+    public List<Role> getAllRoles() {
+        return roleRepository.findAll();
+    }
 }
